requirements.txt
asgiref==3.11.0
Django==6.0.1
django-ckeditor==6.7.3
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.19
django-sass-processor==1.4.2
pillow==12.1.0
sqlparse==0.5.5
typing_extensions==4.15.0
